job_search_platform/job_search_platform/urls.py
"""job_search_platform URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.conf import settings
from django.urls import include, path
from django.contrib import admin
from django.contrib.auth import views as auth_views
from main_platform.views import logoutAndLogin

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', auth_views.LoginView.as_view(redirect_authenticated_user=True), name='login'),
    path('logout', logoutAndLogin,  name='logout'),
    path('main/', include('main_platform.urls',
                          namespace='main_platform')),
]

if settings.DEBUG:
    import debug_toolbar

    urlpatterns += [
        path('__debug__/', include(debug_toolbar.urls)),
    ]
