requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-debug-toolbar==1.11
django-mathfilters==1.0.0
idna==2.8
psycopg2>=2.7
pytz==2018.9
requests==2.21.0
sqlparse==0.2.4
urllib3==1.24.1
whitenoise==4.1.2

selenium~=3.141.0
self~=2020.12.3